Many updates
- hygiene update
- interaction
- shop
- week system
- bug fixes
- controls on pause
