+++ Assets/Models/Database/DrinksItemsDatabase.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad50a9d06845806469e2fd196e77c671
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
