+++ Assets/Models/Database/FoodsItemsDatabase.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 476d95cff5e586048834fc0639a448d1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
