--- Assets/Models/Database/ShopItemsDatabase.asset
@@ -1,50 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 297915caca9a9304e9892f2c25c51c2e, type: 3}
-  m_Name: ShopItemsDatabase
-  m_EditorClassIdentifier: 
-  items:
-  - itemImg: {fileID: 21300000, guid: e80cee7b1c963e44aa18a69dd4e312fd, type: 3}
-    itemName: Milktea
-    itemEnergy: 5
-    price: 23
-    itemType: 0
-    levelRequirement: 0
-    isPurchased: 0
-  - itemImg: {fileID: 21300000, guid: b110e462a2fadb349b1ba39131fea3ba, type: 3}
-    itemName: 'Energy '
-    itemEnergy: 6
-    price: 30
-    itemType: 0
-    levelRequirement: 0
-    isPurchased: 0
-  - itemImg: {fileID: 21300000, guid: 31b230db62895ac4b9f8a03953b15c02, type: 3}
-    itemName: Coffee
-    itemEnergy: 4
-    price: 15
-    itemType: 0
-    levelRequirement: 0
-    isPurchased: 0
-  - itemImg: {fileID: 21300000, guid: a4f85854550ada64c91edaa1effc7e87, type: 3}
-    itemName: Coke
-    itemEnergy: 5
-    price: 23
-    itemType: 0
-    levelRequirement: 0
-    isPurchased: 0
-  - itemImg: {fileID: 21300000, guid: e4a9d63d53c93234a860b3b1a2a2b2b2, type: 3}
-    itemName: Tea
-    itemEnergy: 4
-    price: 16
-    itemType: 0
-    levelRequirement: 0
-    isPurchased: 0
