--- Assets/Models/UI/Hygiene/HygieneSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6beb0aa95ea71f42836a7d05a9dbed7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
