--- Assets/Models/UI/QuizPrefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a0910bd6c32b7df41a2db95145ed301f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
