--- Assets/Scripts/CheatCodes.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 58a8f07e422e37c428a7eddf3bcab0f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
