+++ Assets/Scripts/Core/Clock.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d545b05665050e4fa85115ee69849bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
