--- Assets/Scripts/RoomChecker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f1f85437bf163874d886b6299045b99c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
