+++ Assets/Scripts/Utils/RoomChecker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1f85437bf163874d886b6299045b99c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
